rename: /main -> /browse
