+++ frontend/src/components/Empty.vue
@@ -13,7 +13,7 @@ const props = defineProps({
   <div class="flex flex-col gap-7 w-full h-full justify-center items-center">
     <span class="text-2xl">{{ props.text }}</span>
     <button
-      @click="router.push({ name: 'main' })"
+      @click="router.push({ name: 'browse' })"
       class="btn btn-outline btn-primary"
     >
       On your matchs
