+++ frontend/src/views/BrowseView.vue
